Avoid declaring classes inside other classes, especially methods
