+++ JammaLib/JammaLib.vcxproj.filters
@@ -314,6 +314,12 @@
     <ClInclude Include="src\midi\MidiTimestampMapper.h">
       <Filter>src\midi</Filter>
     </ClInclude>
+    <ClInclude Include="src\midi\MidiIndexedOutputSink.h">
+      <Filter>src\midi</Filter>
+    </ClInclude>
+    <ClInclude Include="src\midi\MidiVstOutputSink.h">
+      <Filter>src\midi</Filter>
+    </ClInclude>
     <ClInclude Include="src\graphics\MidiModel.h">
       <Filter>src\graphics</Filter>
     </ClInclude>
@@ -641,6 +647,12 @@
       <Filter>src\midi</Filter>
     </ClCompile>
     <ClCompile Include="src\midi\MidiTimestampMapper.cpp">
+      <Filter>src\midi</Filter>
+    </ClCompile>
+    <ClCompile Include="src\midi\MidiIndexedOutputSink.cpp">
+      <Filter>src\midi</Filter>
+    </ClCompile>
+    <ClCompile Include="src\midi\MidiVstOutputSink.cpp">
       <Filter>src\midi</Filter>
     </ClCompile>
     <ClCompile Include="src\graphics\MidiModel.cpp">
